requirements.txt
numpy==1.23.1
Pillow==9.2.0
requests~=2.28.1
opencv-python~=4.6.0.66
mss~=6.1.0